Move tests for client.advanced_search() into their own file
